+++ src/go/cmd/aliases.go
@@ -1,0 +1,6 @@
+package cmd
+
+const (
+	deleteAlias  = "del"
+	restartAlias = "res"
+)